+++ frontend/components/queries/QueryResultsTable/QueryResultsTable.jsx
@@ -150,8 +150,8 @@ class QueryResultsTable extends Component {
       campaign,
       onExportQueryResults,
       onExportErrorsResults,
-      isQueryFullScreen,
-      isQueryShrinking,
+      isQueryFullScreen = false,
+      isQueryShrinking = false,
       onToggleQueryFullScreen,
       onRunQuery,
       onStopQuery,
