+++ frontend/pages/queries/QueryPage1/QueryPage.tsx
@@ -1,49 +1,26 @@
 import React, { useState, useEffect } from "react";
-import { Link } from "react-router";
-import classnames from "classnames";
 import { connect, useDispatch } from "react-redux";
 import { useQuery, useMutation } from "react-query";
-import { push } from "react-router-redux";
-import moment from "moment";
-import FileSaver from "file-saver";
-import { filter, isEqual } from "lodash";
 
 // @ts-ignore
 import Fleet from "fleet";
-import { formatSelectedTargetsForApi } from "fleet/helpers";
-import queryAPI from "services/entities/queries";
-import PATHS from "router/paths"; // @ts-ignore
-import debounce from "utilities/debounce"; // @ts-ignore
-import convertToCSV from "utilities/convert_to_csv"; // @ts-ignore
-import deepDifference from "utilities/deep_difference";
+import {
+  selectOsqueryTable, // @ts-ignore
+} from "redux/nodes/components/QueryPages/actions";
+import queryAPI from "services/entities/queries"; // @ts-ignore
 import permissionUtils from "utilities/permissions";
 import { IQueryFormData, IQuery } from "interfaces/query";
-import { ITarget, ITargetsResponse } from "interfaces/target";
+import { ITarget } from "interfaces/target";
 import { IOsqueryTable } from "interfaces/osquery_table";
 import { IUser } from "interfaces/user";
 import { ICampaign } from "interfaces/campaign";
 
 // @ts-ignore
-import { renderFlash } from "redux/nodes/notifications/actions";
-import {
-  selectOsqueryTable,
-  setSelectedTargets,
-  setSelectedTargetsQuery, // @ts-ignore
-} from "redux/nodes/components/QueryPages/actions"; // @ts-ignore
-import campaignHelpers from "redux/nodes/entities/campaigns/helpers"; // @ts-ignore
-
-import QueryForm from "components/forms/queries/QueryForm1"; // @ts-ignore
 import WarningBanner from "components/WarningBanner"; // @ts-ignore
-import QueryPageSelectTargets from "components/queries/QueryPageSelectTargets"; // @ts-ignore
-import QueryResultsTable from "components/queries/QueryResultsTable"; // @ts-ignore
 import QuerySidePanel from "components/side_panels/QuerySidePanel"; // @ts-ignore
-import validateQuery from "components/forms/validators/validate_query";
-import {
-  hasSavePermissions,
-  selectHosts,
-} from "pages/queries/QueryPage1/helpers";
-
-import BackChevron from "../../../../assets/images/icon-chevron-down-9x6@2x.png";
+import QueryEditor from "pages/queries/QueryPage1/components/screens/QueryEditor";
+import SelectTargets from "pages/queries/QueryPage1/components/screens/SelectTargets";
+import RunQuery from "pages/queries/QueryPage1/components/screens/RunQuery";
 
 interface IQueryPageProps {
   queryIdForEdit: string;
@@ -53,20 +30,10 @@ interface IQueryPageProps {
   isBasicTier: boolean;
 }
 
-let runQueryInterval: any = null;
-let globalSocket: any = null;
-let previousSocketData: any = null;
-
 const PAGE_STEP = {
   EDITOR: "EDITOR",
   TARGETS: "TARGETS",
   RUN: "RUN",
-  RESULTS: "RESULTS",
-};
-
-const QUERY_RESULTS_OPTIONS = {
-  FULL_SCREEN: "FULL_SCREEN",
-  SHRINKING: "SHRINKING",
 };
 
 const baseClass = "query-page";
@@ -78,25 +45,14 @@ const QueryPage = ({
   currentUser,
   isBasicTier,
 }: IQueryPageProps) => {
-  const { EDITOR, TARGETS, RUN, RESULTS } = PAGE_STEP;
   const dispatch = useDispatch();
 
-  const [step, setStep] = useState<string>(EDITOR);
+  const [step, setStep] = useState<string>(PAGE_STEP.EDITOR);
   const [typedQueryBody, setTypedQueryBody] = useState<string>("");
-  const [runQueryMilliseconds, setRunQueryMilliseconds] = useState<number>(0);
   const [campaign, setCampaign] = useState<ICampaign | null>(null);
   const [queryIsRunning, setQueryIsRunning] = useState<boolean>(false);
-  const [targetsCount, setTargetsCount] = useState<number>(0);
-  const [targetsError, setTargetsError] = useState<string | null>(null);
-  const [queryResultsToggle, setQueryResultsToggle] = useState<any>(null);
-  const [queryPosition, setQueryPosition] = useState<any>({});
-  const [
-    selectRelatedHostTarget,
-    setSelectRelatedHostTarget,
-  ] = useState<boolean>(true);
-  const [observerShowSql, setObserverShowSql] = useState<boolean>(false);
+  const [showQueryEditor, setShowQueryEditor] = useState<boolean>(false);
   const [liveQueryError, setLiveQueryError] = useState<string>("");
-  const [csvQueryName, setCsvQueryName] = useState<string>("Query Results");
 
   const {
     status,
@@ -129,264 +85,8 @@ const QueryPage = ({
     checkLiveQuery();
   }, []);
 
-  const removeSocket = () => {
-    if (globalSocket) {
-      globalSocket.close();
-      globalSocket = null;
-      previousSocketData = null;
-    }
-
-    return false;
-  };
-
-  const setupDistributedQuery = (socket: any) => {
-    globalSocket = socket;
-    const update = () => {
-      setRunQueryMilliseconds(runQueryMilliseconds + 1000);
-    };
-
-    if (!runQueryInterval) {
-      runQueryInterval = setInterval(update, 1000);
-    }
-
-    return false;
-  };
-
-  const teardownDistributedQuery = () => {
-    if (runQueryInterval) {
-      clearInterval(runQueryInterval);
-      runQueryInterval = null;
-    }
-
-    setQueryIsRunning(false);
-    setRunQueryMilliseconds(0);
-    removeSocket();
-
-    return false;
-  };
-
-  const destroyCampaign = () => {
-    setCampaign(null);
-
-    return false;
-  };
-
-  const onSaveQueryFormSubmit = debounce(async (formData: IQueryFormData) => {
-    try {
-      const { query }: { query: IQuery } = await createQuery(formData);
-      dispatch(push(PATHS.EDIT_QUERY(query)));
-      dispatch(renderFlash("success", "Query created!"));
-    } catch (createError) {
-      console.log(createError);
-      dispatch(
-        renderFlash(
-          "error",
-          "Something went wrong creating your query. Please try again."
-        )
-      );
-    }
-  });
-
-  const onUpdateQuery = async (formData: IQueryFormData) => {
-    if (!storedQuery) {
-      return false;
-    }
-
-    const updatedQuery = deepDifference(formData, storedQuery);
-
-    try {
-      await queryAPI.update(storedQuery, updatedQuery);
-      dispatch(renderFlash("success", "Query updated!"));
-    } catch (updateError) {
-      console.log(updateError);
-      dispatch(
-        renderFlash(
-          "error",
-          "Something went wrong updating your query. Please try again."
-        )
-      );
-    }
-
-    return false;
-  };
-
-  const onChangeQueryFormField = (fieldName: string, value: string) => {
-    if (fieldName === "query") {
-      setTypedQueryBody(value);
-    }
-
-    return false;
-  };
-
   const onOsqueryTableSelect = (tableName: string) => {
     dispatch(selectOsqueryTable(tableName));
-
-    return false;
-  };
-
-  const onRunQuery = debounce(async () => {
-    const sql = typedQueryBody || storedQuery?.query;
-    const { error: validationError } = validateQuery(sql);
-
-    if (!sql) {
-      return false;
-    }
-
-    if (!selectedTargets.length) {
-      setTargetsError("You must select at least one target to run a query");
-
-      return false;
-    }
-
-    if (!targetsCount) {
-      setTargetsError(
-        "You must select a target with at least one host to run a query"
-      );
-
-      return false;
-    }
-
-    if (validationError) {
-      dispatch(renderFlash("error", validationError));
-      return false;
-    }
-
-    const selected = formatSelectedTargetsForApi(selectedTargets);
-
-    removeSocket();
-    destroyCampaign();
-
-    try {
-      const campaignResponse = await queryAPI.run({ query: sql, selected });
-
-      Fleet.websockets.queries.run(campaignResponse.id).then((socket: any) => {
-        setupDistributedQuery(socket);
-        setCampaign(campaignResponse);
-        setQueryIsRunning(true);
-
-        socket.onmessage = ({ data }: any) => {
-          const socketData = JSON.parse(data);
-
-          if (previousSocketData && isEqual(socketData, previousSocketData)) {
-            return false;
-          }
-
-          previousSocketData = socketData;
-
-          const {
-            campaign: socketCampaign,
-            queryIsRunning: socketQueryIsRunning,
-          } = campaignHelpers.updateCampaignState(socketData);
-          socketCampaign && setCampaign(socketCampaign);
-          socketQueryIsRunning !== undefined &&
-            setQueryIsRunning(socketQueryIsRunning);
-
-          if (
-            socketData.type === "status" &&
-            socketData.data.status === "finished"
-          ) {
-            return teardownDistributedQuery();
-          }
-
-          return false;
-        };
-      });
-    } catch (campaignError) {
-      if (campaignError === "resource already created") {
-        dispatch(
-          renderFlash(
-            "error",
-            "A campaign with the provided query text has already been created"
-          )
-        );
-
-        return false;
-      }
-
-      dispatch(renderFlash("error", campaignError));
-      return false;
-    }
-  });
-
-  const onStopQuery = (evt: React.MouseEvent<HTMLButtonElement>) => {
-    evt.preventDefault();
-
-    return teardownDistributedQuery();
-  };
-
-  const onFetchTargets = (
-    targetSearchText: string,
-    targetResponse: ITargetsResponse
-  ) => {
-    const { targets_count: responseTargetsCount } = targetResponse;
-
-    dispatch(setSelectedTargetsQuery(targetSearchText));
-    setTargetsCount(responseTargetsCount);
-
-    return false;
-  };
-
-  const onTargetSelect = (selected: ITarget[]) => {
-    setTargetsError(null);
-    dispatch(setSelectedTargets(selectedTargets));
-
-    return false;
-  };
-
-  const onExportQueryResults = (evt: React.MouseEvent<HTMLButtonElement>) => {
-    evt.preventDefault();
-
-    if (!campaign) {
-      return false;
-    }
-
-    const { query_results: queryResults } = campaign;
-
-    if (queryResults) {
-      const csv = convertToCSV(queryResults, (fields: string[]) => {
-        const result = filter(fields, (f) => f !== "host_hostname");
-        result.unshift("host_hostname");
-
-        return result;
-      });
-
-      const formattedTime = moment(new Date()).format("MM-DD-YY hh-mm-ss");
-      const filename = `${csvQueryName} (${formattedTime}).csv`;
-      const file = new global.window.File([csv], filename, {
-        type: "text/csv",
-      });
-
-      FileSaver.saveAs(file);
-    }
-
-    return false;
-  };
-
-  const onExportErrorsResults = (evt: React.MouseEvent<HTMLButtonElement>) => {
-    evt.preventDefault();
-
-    if (!campaign) {
-      return false;
-    }
-
-    const { errors } = campaign;
-
-    if (errors) {
-      const csv = convertToCSV(errors, (fields: string[]) => {
-        const result = filter(fields, (f) => f !== "host_hostname");
-        result.unshift("host_hostname");
-
-        return result;
-      });
-
-      const formattedTime = moment(new Date()).format("MM-DD-YY hh-mm-ss");
-      const filename = `${csvQueryName} Errors (${formattedTime}).csv`;
-      const file = new global.window.File([csv], filename, {
-        type: "text/csv",
-      });
-
-      FileSaver.saveAs(file);
-    }
 
     return false;
   };
@@ -408,88 +108,55 @@ const QueryPage = ({
     );
   };
 
-  const renderTargetsInput = () => {
-    return (
-      <QueryPageSelectTargets
-        campaign={campaign}
-        error={targetsError}
-        onFetchTargets={onFetchTargets}
-        onRunQuery={onRunQuery}
-        onStopQuery={onStopQuery}
-        onTargetSelect={onTargetSelect}
-        queryIsRunning={queryIsRunning}
-        selectedTargets={selectedTargets}
-        targetsCount={targetsCount}
-        queryTimerMilliseconds={runQueryMilliseconds}
-        disableRun={liveQueryError !== undefined}
-        queryId={queryIdForEdit}
-        isBasicTier={isBasicTier}
-      />
-    );
+  const goToSelectTargets = () => {
+    setStep(PAGE_STEP.TARGETS);
   };
 
-  const renderResultsTable = () => {
-    // const loading = queryIsRunning && !campaign.hosts_count.total;
-    const isQueryFullScreen =
-      queryResultsToggle === QUERY_RESULTS_OPTIONS.FULL_SCREEN;
-    const isQueryShrinking =
-      queryResultsToggle === QUERY_RESULTS_OPTIONS.SHRINKING;
-    const resultsClasses = classnames(`${baseClass}__results`, "body-wrap", {
-      [`${baseClass}__results--loading`]: queryIsRunning,
-      [`${baseClass}__results--full-screen`]: isQueryFullScreen,
-    });
+  const renderScreen = () => {
+    const { TARGETS, RUN } = PAGE_STEP;
 
-    // if (isEqual(campaign, DEFAULT_CAMPAIGN)) {
-    //   return false;
-    // }
-
-    return (
-      <div className={resultsClasses}>
-        <QueryResultsTable
-          campaign={campaign}
-          onExportQueryResults={onExportQueryResults}
-          onExportErrorsResults={onExportErrorsResults}
-          isQueryFullScreen={isQueryFullScreen}
-          isQueryShrinking={isQueryShrinking}
-          // onToggleQueryFullScreen={onToggleQueryFullScreen}
-          onRunQuery={onRunQuery}
-          onStopQuery={onStopQuery}
-          onTargetSelect={onTargetSelect}
-          queryIsRunning={queryIsRunning}
-          queryTimerMilliseconds={runQueryMilliseconds}
-        />
-      </div>
-    );
+    switch (step) {
+      case TARGETS:
+        const step2Opts = {
+          selectedTargets,
+          campaign,
+          isBasicTier,
+          queryIdForEdit,
+          dispatch,
+        };
+        return <SelectTargets {...step2Opts} />;
+      case RUN:
+        const step3Opts = {
+          typedQueryBody,
+          storedQuery,
+          campaign,
+          selectedTargets,
+          queryIsRunning,
+          setQueryIsRunning,
+          setCampaign,
+          dispatch,
+        };
+        return <RunQuery {...step3Opts} />;
+      default:
+        const step1Opts = {
+          baseClass, 
+          currentUser, 
+          dispatch,
+          storedQuery,
+          createQuery,
+          error,
+          goToSelectTargets,
+          setTypedQueryBody,
+        };
+        return <QueryEditor {...step1Opts} />;
+    };
   };
 
   return (
     <div className={`${baseClass} has-sidebar`}>
       <div className={`${baseClass}__content`}>
-        <div className={`${baseClass}__form body-wrap`}>
-          <Link to={PATHS.MANAGE_QUERIES} className={`${baseClass}__back-link`}>
-            <img src={BackChevron} alt="back chevron" id="back-chevron" />
-            <span>Back to queries</span>
-          </Link>
-          <QueryForm
-            onCreateQuery={onSaveQueryFormSubmit}
-            onChangeFunc={onChangeQueryFormField}
-            onOsqueryTableSelect={onOsqueryTableSelect}
-            onRunQuery={onRunQuery}
-            onStopQuery={onStopQuery}
-            onUpdate={onUpdateQuery}
-            queryIsRunning={queryIsRunning}
-            serverErrors={error || {}}
-            selectedOsqueryTable={selectedOsqueryTable}
-            title={storedQuery?.name || "New query"}
-            hasSavePermissions={hasSavePermissions(currentUser)}
-          />
-        </div>
+        {renderScreen()}
         {renderLiveQueryWarning()}
-        {/* ONLY SHOW FOR STEP 2 */}
-        {/* {renderTargetsInput()} */}
-
-        {/* ONLY SHOW FOR STEP 3 */}
-        {/* {renderResultsTable()} */}
       </div>
       <QuerySidePanel
         onOsqueryTableSelect={onOsqueryTableSelect}
